🐛 (#2097): add support to autolink
auto link is included in original tiptap extension-link.

We should keep the original behavior as much as possible.
