+++ src/marks/Link.js
@@ -20,11 +20,9 @@
  *
  */
 
-import { generateUrl } from '@nextcloud/router'
 import TipTapLink from '@tiptap/extension-link'
-import { Plugin } from 'prosemirror-state'
 import { domHref, parseHref } from './../helpers/links'
-import markdownit from './../markdownit'
+import { clickHandler } from '../plugins/link'
 
 const Link = TipTapLink.extend({
 
@@ -53,53 +51,18 @@ const Link = TipTapLink.extend({
 	}, 0],
 
 	addProseMirrorPlugins() {
+		const plugins = this.parent()
+			// remove original handle click
+			.filter(({ key }) => {
+				return !key.startsWith('handleClickLink')
+			})
+
 		if (!this.options.openOnClick) {
-			return []
+			return plugins
 		}
-		return [
-			new Plugin({
-				props: {
-					handleClick: (view, pos, event) => {
-						const attrs = this.editor.getAttributes('link')
 
-						const isLink = event.target instanceof HTMLAnchorElement || event.target.parentElement instanceof HTMLAnchorElement
-						if (attrs.href && isLink) {
-							const linkElement = event.target.parentElement instanceof HTMLAnchorElement ? event.target.parentElement : event.target
-							event.stopPropagation()
-							const htmlHref = linkElement.href
-							if (event.button === 0 && !event.ctrlKey && htmlHref.startsWith(window.location.origin)) {
-								const query = OC.parseQueryString(htmlHref)
-								const fragment = OC.parseQueryString(htmlHref.split('#').pop())
-								if (query.dir && fragment.relPath) {
-									const filename = fragment.relPath.split('/').pop()
-									const path = `${query.dir}/${filename}`
-									document.title = `${filename} - ${OC.theme.title}`
-									if (window.location.pathname.match(/apps\/files\/$/)) {
-										// The files app still lacks a popState handler
-										// to allow for using the back button
-										// OC.Util.History.pushState('', htmlHref)
-									}
-									OCA.Viewer.open({ path })
-									return
-								}
-								if (query.fileId) {
-									// open the direct file link
-									window.open(generateUrl(`/f/${query.fileId}`))
-									return
-								}
-							}
-
-							if (!markdownit.validateLink(htmlHref)) {
-								console.error('Invalid link', htmlHref)
-								return
-							}
-
-							window.open(htmlHref)
-						}
-					},
-				},
-			}),
-		]
+		// add custom click handler
+		return [...plugins, clickHandler({ editor: this.editor, type: this.type })]
 	},
 })
